Split `PurchasesTests` step 4: extracted `PurchasesRestoreTests`
Follow up to #1609.
